requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cookiecutter==2.1.1
coverage==6.5.0
cssselect==1.1.0
cssutils==2.6.0
Deprecated==1.2.13
Django==4.1.3
django-admin-honeypot==1.1.0
django-anymail==8.6
django-appconf==1.0.5
django-autoslug==1.9.8
django-bootstrap4==22.2
django-bootstrap5==22.1
django-classy-tags==3.0.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-filter==22.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-redis==5.2.0
django-redis-cache==3.0.1
django-render-block==0.9.2
django-sekizai==4.0.0
django-templated-email==3.0.0
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
flake8==5.0.4
h11==0.14.0
hiredis==2.0.0
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
Jinja2==3.1.2
jinja2-time==0.2.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
outcome==1.2.0
packaging==21.3
pathspec==0.10.1
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
premailer==3.10.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-memcached==1.59
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
redis==3.5.3
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.9.0
